user/my api에 themeExp 정보 추가
