--- src/main/java/com/fours/tolevelup/controller/Request/Request.java
@@ -1,16 +1,0 @@
-package com.fours.tolevelup.controller.Request;
-
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class Request {
-
-    private String s;
-
-    public static void UserJoin(){
-
-    }
-
-
-}
