+++ src/main/java/com/fours/tolevelup/controller/request/UserRequest.java
@@ -1,4 +1,4 @@
-package com.fours.tolevelup.controller.Request;
+package com.fours.tolevelup.controller.request;
 
 
 import lombok.AllArgsConstructor;
