+++ src/main/java/com/fours/tolevelup/controller/response/UserResponse.java
@@ -1,12 +1,16 @@
-package com.fours.tolevelup.controller.Response;
+package com.fours.tolevelup.controller.response;
 
 
+import com.fours.tolevelup.model.ThemeExpDTO;
 import com.fours.tolevelup.model.UserRole;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+import javax.inject.Inject;
 import java.sql.Date;
+import java.util.List;
 
 @Getter
 @AllArgsConstructor
@@ -38,7 +42,7 @@ public class UserResponse {
         private String intro;
         private UserRole role;
         private Date registeredAt;
-
+        private List<ThemeExpDTO.user> themeExp;
     }
 
 }
