+++ src/main/java/com/fours/tolevelup/exception/CustomAuthenticationEntryPoint.java
@@ -1,6 +1,6 @@
 package com.fours.tolevelup.exception;
 
-import com.fours.tolevelup.controller.Response.Response;
+import com.fours.tolevelup.controller.response.Response;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
