+++ src/main/java/com/fours/tolevelup/exception/GlobalControllerAdvice.java
@@ -1,7 +1,7 @@
 package com.fours.tolevelup.exception;
 
 
-import com.fours.tolevelup.controller.Response.Response;
+import com.fours.tolevelup.controller.response.Response;
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.http.HttpStatus;
