+++ src/main/java/com/fours/tolevelup/service/themeexp/ThemeExpService.java
@@ -1,12 +1,13 @@
 package com.fours.tolevelup.service.themeexp;
 
+import com.fours.tolevelup.model.ThemeExpDTO;
 import com.fours.tolevelup.model.entity.Mission;
 
 import java.util.List;
 
 public interface ThemeExpService {
 
-    List<ThemeExpDTO.ThemeExp> findUserThemeExps(String id);
+    //List<ThemeExpDTO.ThemeExp> findUserThemeExps(String id);
 
     void plusUserThemeExp(String user_id,Mission mission);
     void minusUserThemeExp(String user_id,Mission mission);
